requirements.txt
# Azure Agentic AI Support Bot Dependencies

# Core AI Framework
semantic-kernel>=1.37.0
openai>=1.3.0
azure-ai-openai>=1.0.0

# Azure SDK
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.1
azure-mgmt-compute>=29.5.0
azure-mgmt-network>=23.1.0
azure-mgmt-monitor>=6.0.1
azure-mgmt-cosmosdb>=9.4.0
azure-cosmos>=4.5.1
azure-mgmt-keyvault>=10.2.3
azure-keyvault-secrets>=4.7.0

# Bot Framework
botbuilder-core>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.9.0

# Azure Functions
azure-functions>=1.18.0
azure-functions-worker>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Security
cryptography>=41.0.8
pyjwt>=2.8.0

# Monitoring
opencensus-ext-azure>=1.1.13
applicationinsights>=0.11.10
